Update gemspec for new gem name and apache license
This updates the gemspec for the new gem name
(topological_inventory-core) and changes to an Apache 2.0 license
consistent with the ManageIQ project.
