--- config/routes.rb
@@ -1,2 +1,0 @@
-TopologicalInventory::Engine.routes.draw do
-end
