+++ lib/topological_inventory-core.rb
@@ -1,0 +1,6 @@
+require "topological_inventory/core/engine"
+
+module TopologicalInventory
+  module Core
+  end
+end
