--- lib/topological_inventory.rb
@@ -1,5 +1,0 @@
-require "topological_inventory/engine"
-
-module TopologicalInventory
-  # Your code goes here...
-end
