+++ lib/topological_inventory/core/version.rb
@@ -1,0 +1,5 @@
+module TopologicalInventory
+  module Core
+    VERSION = '0.1.0'
+  end
+end
