--- lib/topological_inventory/version.rb
@@ -1,3 +1,0 @@
-module TopologicalInventory
-  VERSION = '0.1.0'
-end
